♻️refactor: 패키지 변경 및 OAuth2 유저 loginId를 email로 설정
